+++ src/main/java/adnyre/model/BaseEntity.java
@@ -1,6 +1,6 @@
 package adnyre.model;
 
 public interface BaseEntity {
-    Long getId();
-    void setId(Long id);
+    int getId();
+    void setId(int id);
 }
